+++ src/app/admin/pages/product/product.component.ts
@@ -47,6 +47,10 @@ constructor(private productService:ProductService,private router:Router) {
   }
 
   delete(id:number){
+
+    const isConfirmed = window.confirm('Silmek istediğinize emin misiniz?');
+    // Kullanıcı "Evet" derse işlemi devam ettir
+  if (isConfirmed) {
     this.productService.delete(id).subscribe({
       next:(x)=>{ console.log(x.body);
         console.log(x.headers);
@@ -68,6 +72,7 @@ constructor(private productService:ProductService,private router:Router) {
 
     })
   }
+}
